% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert a data.frame to a distance matrix}
\usage{
df_to_matrix(data, lower = TRUE, upper = TRUE, diag = TRUE, na_to_zero = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} with the following three columns: \code{from} (the
first site), \code{to} (the second site) and \code{distance} (the value of the link
between two sites). The output of the function \code{\link[=distance_along]{distance_along()}}.}

\item{lower}{a \code{logical} value. If \code{TRUE} (default), keep values in the
lower triangle of the matrix. Otherwise they will be replaced by \code{NA}
(or \code{0}).}

\item{upper}{a \code{logical} value. If \code{TRUE} (default), keep values in the
upper triangle of the matrix. Otherwise they will be replaced by \code{NA}
(or \code{0}).}

\item{diag}{a \code{logical} value. If \code{TRUE} (default), keep values in the
diagonal of the matrix. Otherwise they will be replaced by \code{NA}
(or \code{0}).}

\item{na_to_zero}{a \code{logical} value. If \code{TRUE} (default), missing edges are
coded as \code{0}. Otherwise they will be coded as \code{NA}.}
}
\value{
A matrix of dimensions \verb{m x n}, where \code{m} is the number of sites
(\code{from}) and \code{n} is the number of sites (\code{to}).
}
\description{
Converts a \code{data.frame} to a distance \code{matrix}.
}
\examples{
# Add an example ----
}
