% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_along.R
\name{distance_along}
\alias{distance_along}
\title{Compute the distance between two locations along a linear shape}
\usage{
distance_along(
  sites,
  along,
  density = 0.01,
  type = "regular",
  mc.cores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{sites}{an \code{sf} object of type \code{POINT}. A spatial object containing
coordinates of sites. Note that the first column must be the
site identifier.}

\item{along}{an \code{sf} object of type \code{LINESTRING}. A spatial object
containing coordinates of the linear shape (e.g. a river, a road, etc.)
to follow while computing distances between the two sites.}

\item{density}{a \code{numeric} of length 1. The density of points to sample on
the linear structure. See \code{\link[sf:st_line_sample]{sf::st_line_sample()}} for further detail.
Default is \code{1/100}.}

\item{type}{a \code{character} of length 1. The method to sample points on the
linear shape. Either \code{regular} (default) or \code{random}.}

\item{mc.cores}{an \code{integer} of length 1. The number of cores to use
(must be lesser than the number of cores available on the machine obtained
with \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{...}{other argument to pass to  \code{\link[sf:st_line_sample]{sf::st_line_sample()}}.}
}
\value{
A three-column \code{data.frame} with:
\itemize{
\item \code{from}, the first site
\item \code{to}, the second site
\item \code{weight}, the distance between the two sites along the linear shape
}
}
\description{
Computes distance between two sites along a linear shape (e.g. a river, a
road, etc.). This function uses parallel computing to speed up the distance
computation.
}
\examples{
## Add an example ----
}
