% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakedist-package.R
\docType{package}
\name{snakedist-package}
\alias{snakedist}
\alias{snakedist-package}
\title{snakedist: Compute Distances along a Linear Shape}
\description{
Computes distance between pair of points along a linear shape (e.g. a river, a road, etc.). This function uses parallel computing to speed up the distance computation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/snakedist}
  \item Report bugs at \url{https://github.com/frbcesab/snakedist/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
